a new program
this is a simple programm that is supposed to show how to work with files but what it actually teachers that's interesting is loop control. The 'continue' function is VERY interesting. Because for whatever reason it seems to reflect the rest of the loop and not just the loop it's in. which means I might be able to use it for loop control instead of having to have 4 different flags.