finalized
finally fixed it the solution was very simple and elegant. i was worried I'd have to turn three test cases and two more flags. When in reality I just had to realize I could return early. Once I realized I wanted to just exit the function and not just the loop it took me stupid long to realize the command for that was starring me in the face.  That's the kind of facepalm that sticks with you.. learning.