change number works
was having issues with being able to change the variables and everyone I went to for help sucks but I figured it out on my own which is fine for me but speaks poorly of the type of help i can expect in the future.

Also learned about string formatting so %s ing all over the place.